requirements.txt
django<2
psycopg2
uwsgi
djangorestframework
django-filter
requests
